Put classes Boogie.Helper, Boogie.TokenTextWriter and Boogie.ListComparer into separate files
